requirements.txt
h5py==2.7.1
matplotlib==2.2.2
numpy==1.14.2
pandas==0.22.0
Pillow==5.0.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
sklearn==0.0
torch==0.3.1
termcolor==1.1.0
cffi==1.11.5
torchfile==0.1.0
torchvision==0.2.0
visdom==0.1.7